Move Krylov dep to extensions
